+++ My_Smart_Factory/Models/Insp/InspEquipSettingRecordModel.cs
@@ -12,7 +12,7 @@ namespace My_Smart_Factory.Models.Insp
         public virtual InspEquipModel? InspEquip { get; set; }      // 검사에 사용한 장비
         public virtual InspSpecModel? InspSpec { get; set; }        // 검사에 기준이 된 스펙
         public DateTime? InspectionDateTime { get; set; }           // 검사 일시
-        public double? IES { get; set; }                        // Inspection Equipment Settings
-        public double? Accuracy { get; set; }                   // 일치율
+        public decimal? IES { get; set; }                        // Inspection Equipment Settings
+        public decimal? Accuracy { get; set; }                   // 일치율
     }
 }
